Exclude TLS1.3 support if older go version is used (#15543)
Only compile with TLS1.3 support for go1.13 or newer is used. If an
older go version is used we stick with TLS1.2 max.

The change also introduces TLSVersionMin/Max and
TLSVersionDefaultMin/Max constants, so to keep the tests intact.
