+++ libbeat/common/transport/tlscommon/tls_test.go
@@ -131,8 +131,8 @@ func TestApplyEmptyConfig(t *testing.T) {
 	}
 
 	cfg := tmp.BuildModuleConfig("")
-	assert.Equal(t, int(tls.VersionTLS11), int(cfg.MinVersion))
-	assert.Equal(t, int(tls.VersionTLS13), int(cfg.MaxVersion))
+	assert.Equal(t, int(TLSVersionDefaultMin), int(cfg.MinVersion))
+	assert.Equal(t, int(TLSVersionDefaultMax), int(cfg.MaxVersion))
 	assert.Len(t, cfg.Certificates, 0)
 	assert.Nil(t, cfg.RootCAs)
 	assert.Equal(t, false, cfg.InsecureSkipVerify)
@@ -163,8 +163,8 @@ func TestApplyWithConfig(t *testing.T) {
 	assert.NotNil(t, cfg.RootCAs)
 	assert.Equal(t, true, cfg.InsecureSkipVerify)
 	assert.Len(t, cfg.CipherSuites, 2)
-	assert.Equal(t, int(tls.VersionTLS11), int(cfg.MinVersion))
-	assert.Equal(t, int(tls.VersionTLS13), int(cfg.MaxVersion))
+	assert.Equal(t, int(TLSVersionDefaultMin), int(cfg.MinVersion))
+	assert.Equal(t, int(TLSVersionDefaultMax), int(cfg.MaxVersion))
 	assert.Len(t, cfg.CurvePreferences, 1)
 	assert.Equal(t, tls.RenegotiateOnceAsClient, cfg.Renegotiation)
 }
@@ -188,8 +188,8 @@ func TestServerConfigDefaults(t *testing.T) {
 		assert.Len(t, cfg.CurvePreferences, 0)
 		// values set by default
 		assert.Equal(t, false, cfg.InsecureSkipVerify)
-		assert.Equal(t, int(tls.VersionTLS11), int(cfg.MinVersion))
-		assert.Equal(t, int(tls.VersionTLS13), int(cfg.MaxVersion))
+		assert.Equal(t, int(TLSVersionDefaultMin), int(cfg.MinVersion))
+		assert.Equal(t, int(TLSVersionDefaultMax), int(cfg.MaxVersion))
 		assert.Equal(t, tls.NoClientCert, cfg.ClientAuth)
 	})
 	t.Run("when CA is explicitly set", func(t *testing.T) {
@@ -214,8 +214,8 @@ func TestServerConfigDefaults(t *testing.T) {
 		assert.Len(t, cfg.CurvePreferences, 0)
 		// values set by default
 		assert.Equal(t, false, cfg.InsecureSkipVerify)
-		assert.Equal(t, int(tls.VersionTLS11), int(cfg.MinVersion))
-		assert.Equal(t, int(tls.VersionTLS13), int(cfg.MaxVersion))
+		assert.Equal(t, int(TLSVersionDefaultMin), int(cfg.MinVersion))
+		assert.Equal(t, int(TLSVersionDefaultMax), int(cfg.MaxVersion))
 		assert.Equal(t, tls.RequireAndVerifyClientCert, cfg.ClientAuth)
 	})
 }
@@ -227,7 +227,6 @@ func TestApplyWithServerConfig(t *testing.T) {
     certificate_authorities: [ca_test.pem]
     verification_mode: none
     client_authentication: optional
-    supported_protocols: [TLSv1.1, TLSv1.2, TLSv1.3]
     cipher_suites:
       - "ECDHE-ECDSA-AES-256-CBC-SHA"
       - "ECDHE-ECDSA-AES-256-GCM-SHA384"
@@ -235,6 +234,10 @@ func TestApplyWithServerConfig(t *testing.T) {
   `
 	var c ServerConfig
 	config, err := common.NewConfigWithYAML([]byte(yamlStr), "")
+	for i, ver := range TLSDefaultVersions {
+		config.SetString("supported_protocols", i, ver.String())
+	}
+
 	if !assert.NoError(t, err) {
 		return
 	}
@@ -254,8 +257,8 @@ func TestApplyWithServerConfig(t *testing.T) {
 	assert.NotNil(t, cfg.ClientCAs)
 	assert.Equal(t, true, cfg.InsecureSkipVerify)
 	assert.Len(t, cfg.CipherSuites, 2)
-	assert.Equal(t, int(tls.VersionTLS11), int(cfg.MinVersion))
-	assert.Equal(t, int(tls.VersionTLS13), int(cfg.MaxVersion))
+	assert.Equal(t, int(TLSVersionDefaultMin), int(cfg.MinVersion))
+	assert.Equal(t, int(TLSVersionDefaultMax), int(cfg.MaxVersion))
 	assert.Len(t, cfg.CurvePreferences, 1)
 	assert.Equal(t, tls.VerifyClientCertIfGiven, cfg.ClientAuth)
 }
