+++ libbeat/common/transport/tlscommon/types.go
@@ -99,25 +99,6 @@ var tlsRenegotiationSupportTypes = map[string]tlsRenegotiationSupport{
 	"freely": tlsRenegotiationSupport(tls.RenegotiateFreelyAsClient),
 }
 
-// TLSVersion type for TLS version.
-type TLSVersion uint16
-
-// Define all the possible TLS version.
-const (
-	TLSVersionSSL30 TLSVersion = tls.VersionSSL30
-	TLSVersion10    TLSVersion = tls.VersionTLS10
-	TLSVersion11    TLSVersion = tls.VersionTLS11
-	TLSVersion12    TLSVersion = tls.VersionTLS12
-	TLSVersion13    TLSVersion = tls.VersionTLS13
-)
-
-// TLSDefaultVersions list of versions of TLS we should support.
-var TLSDefaultVersions = []TLSVersion{
-	TLSVersion11,
-	TLSVersion12,
-	TLSVersion13,
-}
-
 type tlsClientAuth int
 
 const (
@@ -130,24 +111,6 @@ var tlsClientAuthTypes = map[string]tlsClientAuth{
 	"none":     tlsClientAuthNone,
 	"optional": tlsClientAuthOptional,
 	"required": tlsClientAuthRequired,
-}
-
-var tlsProtocolVersions = map[string]TLSVersion{
-	"SSLv3":   TLSVersionSSL30,
-	"SSLv3.0": TLSVersionSSL30,
-	"TLSv1":   TLSVersion10,
-	"TLSv1.0": TLSVersion10,
-	"TLSv1.1": TLSVersion11,
-	"TLSv1.2": TLSVersion12,
-	"TLSv1.3": TLSVersion13,
-}
-
-var tlsProtocolVersionsInverse = map[TLSVersion]string{
-	TLSVersionSSL30: "SSLv3",
-	TLSVersion10:    "TLSv1.0",
-	TLSVersion11:    "TLSv1.1",
-	TLSVersion12:    "TLSv1.2",
-	TLSVersion13:    "TLSv1.3",
 }
 
 // TLSVerificationMode represents the type of verification to do on the remote host,
@@ -165,24 +128,6 @@ const (
 	//       postVerifyTLSConnection
 	// VerifyCertificate
 )
-
-func (v TLSVersion) String() string {
-	if s, ok := tlsProtocolVersionsInverse[v]; ok {
-		return s
-	}
-	return "unknown"
-}
-
-//Unpack transforms the string into a constant.
-func (v *TLSVersion) Unpack(s string) error {
-	version, found := tlsProtocolVersions[s]
-	if !found {
-		return fmt.Errorf("invalid tls version '%v'", s)
-	}
-
-	*v = version
-	return nil
-}
 
 var tlsVerificationModes = map[string]TLSVerificationMode{
 	"":     VerifyFull,
