+++ libbeat/outputs/transport/tls.go
@@ -35,20 +35,6 @@ type TLSConfig = tlscommon.TLSConfig
 // TLSVersion type for TLS version.
 type TLSVersion = tlscommon.TLSVersion
 
-// Define all the possible TLS version.
-const (
-	TLSVersionSSL30 = tlscommon.TLSVersionSSL30
-	TLSVersion10    = tlscommon.TLSVersion10
-	TLSVersion11    = tlscommon.TLSVersion11
-	TLSVersion12    = tlscommon.TLSVersion12
-)
-
-// Constants of the supported verification mode.
-const (
-	VerifyFull = tlscommon.VerifyFull
-	VerifyNone = tlscommon.VerifyNone
-)
-
 func TLSDialer(forward Dialer, config *TLSConfig, timeout time.Duration) (Dialer, error) {
 	return TestTLSDialer(testing.NullDriver, forward, config, timeout)
 }
